% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2d.R
\name{slr}
\alias{slr}
\title{Simple Linear Regression (internal)}
\usage{
slr(dt, .x, .y, .z)
}
\arguments{
\item{dt}{a data.table object}

\item{.x}{predictor varaible}

\item{.y}{response variable}

\item{.z}{if grouping needs to be done}
}
\value{
data.table object
}
\description{
With data.table Grouping
}
