% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{hist2d}
\alias{hist2d}
\title{2-Dimensional Histogram}
\usage{
hist2d(dt, x, y, xLab, yLab, xLim, yLim, facet = 0, z = NULL, binWidth,
  hasLine = F, palette = "Reds")
}
\arguments{
\item{dt}{Data in the class of \code{\link[data.table]{data.table}}.
Do not name column variable with \code{.x} or \code{.y}.}

\item{x, y}{The \emph{name} of the columns in the \code{dt} that will be used as coordinates of the 2D histogram.
The columns must be of numeric vectors.
These are \code{(x,y)} coordinates when constructing a scatter plot.
Behind the scene,
If \code{x} and/or \code{y}  is not supplied, the first and/or second column of \code{dt} will be used.}

\item{xLab, yLab}{Character strings to be used as axis labels.
If not supplied, the value(s) of \code{x} or \code{y} will be used.}

\item{xLim, yLim}{The limits of coordinates of x or y axis that will be shown in the plot.
It does \emph{not} affect the regression line if \code{hasLine} is \code{TRUE}.
If not supplied, The values will be the minimum and the maximum value of corresponding dimension.}

\item{facet}{Any of \code{{0,1,2}}.
It refers to the \pkg{ggplot2}'s faceting technique.
Multiple 2D histograms will be displayed as a sequence of panels which enables grouping comparison.
The value represents the number of additional categorical variable(s).}

\item{z}{A vector of column name(s) of \code{dt} that will be used for faceting.
The length of the vector must match the number in \code{facet}.
Please, set it to \code{NULL} when \code{facet==0}.}

\item{binWidth}{If not supplied, Freedman and Diaconis’s rule is applied to each dimension.}

\item{hasLine}{logical with \code{FALSE} as the default.
If \code{TRUE}, draw and annotate a simple linear regression line of \code{y} against \code{x}}
}
\description{
2D Histogram is an alternative to traditional scatter plot.
Similar with histogram, it constructs bins of regular size, and count the number of observations found in each bin.
However, 2 dimensions are involved in the binning.
It also use colour gradient instead of bar height to represent the count number.
This plotting technique will be more useful when there are points which are closed or even stacked in the scatter plot.
}
